% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{BTSinFLRM}
\alias{BTSinFLRM-package}
\alias{BTSinFLRM}
\title{Bootstrap inference in functional linear regression models}
\description{
The package implements the residual, paired, and (mutliplier) wild bootstrap methods described in Yeon, Dai, and Nordman (2023, 2024, 2025);
these are developed for inference on both mean response and projections of the slope function at new predictors/regressors 
in functional linear regression models (FLRMs) with scalar response.
The residual bootstrap is designed only for homoscedastic FLRMs
while the paired and wild bootstraps are developed to deal with heteroscedastic error cases.
It also implements bootstrap two-sample tests for scalar-on-function regression based on the work by Yeon and Kokoszka (2025). 

\code{RBinFLRM} provides confidence and prediction intervals for projections 
based on residual bootstrap proposed by Yeon, Dai, and Nordman (2023).
\code{PBinFLRM} performs a modified paired bootstrap method by Yeon, Dai, and Nordman (2024)
to find confidence intervals for projections
and to compute p-values from the hypothesis tests 
of whether the projections onto multiple new regressors are simultaneously zero or not.
\code{WBinFLRM} shows the same results as the ones from \code{PBinFLRM} 
but based on (multiplier) wild bootstrap developed in Yeon, Dai, and Nordman (2025).

\code{test2sofr} conducts hypothesis testing for two-sample scalar-on-function linear regression developed by Yeon and Kokoszka (2025).
The test statistics employ either pooled or simultaneous diagonalization
and their asymptotic chi-squared distributions are approximated by residual-type two-sample bootstrap methods
either imposing the null or not. 

For more details see the help vignette:
\code{vignette("BTSinFLRMvign", package = "BTSinFLRM")}.
}
\examples{
vignette("BTSinFLRMvign", package = "BTSinFLRM")
}
\references{
Yeon, H., Dai, X., and Nordman, D. (2023). Bootstrap inference in functional linear regression models. \emph{Bernoulli}

Yeon, H., Dai, X., and Nordman, D. (2024). Bootstrap inference in functional linear regression models under heteroscedasticity. \emph{Electronic Journal of Statistics}
    
Yeon, H., Dai, X., and Nordman, D. (2025). Wild bootstrap for mean response inference in functional linear regression models. \emph{Submitted}

Yeon, H. and Kokoszka, P. (2025). Bootstrap two-sample tests for scalar-on-function regression. \emph{Submitted}
}
\seealso{
\code{\link{RBinFLRM}}
\code{\link{PBinFLRM}}
\code{\link{WBinFLRM}}
\code{\link{test2sofr}}
}
\author{
Hyemin Yeon \email{hyeon1@kent.edu}
}
