% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{BTSinFLRM}
\alias{BTSinFLRM}
\title{Bootstrap inference in functional linear regression models with scalar response.}
\description{
The package implements the residual, paired, and (mutliplier) wild bootstrap methods described in Yeon, Dai, and Nordman (2023, 2024a, 2024b);
these are developed for inference on projections of the slope function onto new predictors/regressors 
(i.e., the mean responses at the new regressors) 
in functional linear regression models (FLRMs) with scalar response.
The residual bootstrap is designed only for homoscedastic FLRMs
while the paired and wild bootstraps are developed to deal with heteroscedastic error cases.
\code{RBinFLRM} provides confidence and prediction intervals for projections 
based on residual bootstrap proposed by Yeon, Dai, and Nordman (2023).
\code{PBinFLRM} performs a modified paired bootstrap method by Yeon, Dai, and Nordman (2024a)
to find confidence intervals for projections
and to compute p-values from the hypothesis tests 
of whether the projections onto multiple new regressors are simultaneously zero or not.
\code{WBinFLRM} shows the same results as the ones from \code{PBinFLRM} 
but based on (multiplier) wild bootstrap developed in Yeon, Dai, and Nordman (2024b).
For more details see the help vignette:
\code{vignette("BTSinFLRMvign", package = "BTSinFLRM")}.
}
\examples{
vignette("BTSinFLRMvign", package = "BTSinFLRM")

}
\references{
Yeon, H., Dai, X., and Nordman, D. (2023). Bootstrap inference in functional linear regression models. To appear in Bernoulli

Yeon, H., Dai, X., and Nordman, D. (2024a). Bootstrap inference in functional linear regression models under heteroscedasticity. Submitted

Yeon, H., Dai, X., and Nordman, D. (2024b). Different bootstrap methods in functional linear regression models. In progress
}
\seealso{
\code{\link{RBinFLRM}}
\code{\link{PBinFLRM}}
\code{\link{WBinFLRM}}
}
\author{
Hyemin Yeon \email{luckyhm@iastate.edu}
}
